+++ jdaviz/configs/imviz/tests/test_viewers.py
@@ -1,5 +1,9 @@
+import numpy as np
 import pytest
 
+from jdaviz.app import Application
+from jdaviz.core.config import get_configuration
+from jdaviz.configs.imviz.helper import Imviz
 from jdaviz.configs.imviz.plugins.viewers import ImvizImageView
 
 
@@ -28,3 +32,24 @@ def test_destroy_viewer_invalid(imviz_app):
     with pytest.raises(ValueError, match='cannot be destroyed'):
         imviz_app.destroy_viewer('imviz-0')
     assert imviz_app.app.get_viewer_ids() == ['imviz-0']
+
+
+def test_mastviz_config():
+    """Use case from https://github.com/spacetelescope/jdaviz/issues/1037"""
+
+    # create a MAST config dict
+    cc = get_configuration('imviz')
+    cc['settings']['viewer_spec'] = cc['settings'].get('configuration', 'default')
+    cc['settings']['configuration'] = 'mastviz'
+    cc['settings']['visible'] = {'menu_bar': False, 'toolbar': False, 'tray': False,
+                                 'tab_headers': False}
+    cc['toolbar'].remove('g-data-tools') if cc['toolbar'].count('g-data-tools') else None
+    cc['toolbar'].remove('g-viewer-creator') if cc['toolbar'].count('g-viewer-creator') else None
+    cc['toolbar'].remove('g-image-viewer-creator') if cc['toolbar'].count('g-image-viewer-creator') else None  # noqa
+
+    app = Application(cc)
+    im = Imviz(app)
+    im.load_data(np.ones((2, 2)), data_label='my_array')
+
+    assert im.app.get_viewer_ids() == ['mastviz-0']
+    assert im.app.data_collection[0].shape == (2, 2)